switch roles to apps since roles is overloaded in puppet
Roles is a great name, except that we use role to mean stuff in puppet, and it does not match 1-1 with what applications are deployed onto that machine, i.e. jetty-server is a role for all jetty servers regardless of the app applied.
Instead we've gone for apps, which is a fact that indicates which apps should be installed