Handle JValues on right of data, not just strings
So in json {"a": "b"} is just as valid as {"a": [1, 2]}
Therefore data isn't a map of String -> String, it's a map of String -> JValue.
But we know what type we actually expect, so data("port") should always be a string, so provide stringData(key) to encode that assumption.
I suspect that this could be nicer, but I'm not sure how right now.