Added a HealthcheckGrace task.
This task introduces a pause when deploying in order to work around an ongoing problem whereby AWS initially reports incoming instances as InService before it correctly identifies them as OutOfService.

It is a bit of a race condition,  but when it occurs it leads to about a minute of downtime (in PROD) during deploys.

 We are not the only people this is happening to, there is a thread over here https://forums.aws.amazon.com/thread.jspa?messageID=457489