Add underlying functionality to trigger a deploy based on a tag
This adds the functionality for the TeamCity integration to detect newly tagged builds and also the
continuous deployment enhancement to register a listener for new tags and act on notifications when
they are delivered.

This is hard to achieve as the TeamCity API is less than amazing. Significant chunks of
code have been re-written in order to avoid repetition whilst keeping a reasonable level
of maintainability.