Make an attempt to kill off long running deployinfo processes
This will now use Process.destroy to kill off a deployinfo process
that is deemed to have locked up in some way. This sends a TERM to
the child process that was started, but doesn't deal with any
children that the process may have started itself.

There are some new configuration properties around this.

Also updated release notes.