Avoid race condition when previewing
Under load, send might not have completed when the preview content template attempts to render resulting in an exception.  I've seen this happen a number of times.  By blocking on the update of the available previews we can avoid this.  There are definitely more heinous examples of blocking around in RiffRaff, so I'm not too worried about this one.