Support stage-specific params for cfn deployment types
Introduces a new property `templateStageParameters` which supports
stage-specific parameters.

Riff Raff will look for a key in the property's map matching the stage
of the deployment. If found, it will merge the contained parameters into
`templateParameters` (the global parameters if you like), with
stage-specific keys overriding global ones in case of collision.