Replace the S3Upload task with a new S3UploadV2 task
The `S3Upload` task does a little too much internally which makes it hard to reason about and hard to use when you want to reliably work out where it has uploaded files. The replacement V2 doesn't do as much magic and moves the responsibility of mapping files to keys to the callee.

At the moment I've only replaced the use of it in the lambda upload action, but would like to completely replace and remove `S3Upload`.