Consistently display timestamps in London time, UK locale
Date formatting was quite inconsistent across the app, e.g.:

* Deploy start time was in London time (actually browser locale's time, I think) on the history page, but the deploy logs were in UTC.
* API keys, etc. had their dates formatted in US style (mm/dd/yy)

There is an argument for displaying all timestamps in UTC, but given
that nearly all users are in London, I think London time is more
convenient. As long as we are consistent (which we now should be), it
doesn't really matter which one we pick.