Make the DeployRunState into an Actor to simplify the state management
The updating of the DeployRunState instances was becoming a messy affair so it has now been changed into an Actor called DeployRunner. The factories are non-trivial to make testing possible but otherwise this seems far more elegant - especially once the stop flag map is converted into an agent that is passed around and used safely by many different parties.