Switch to ValidatedNel / Validated instead of Either
This makes much of the code a bit cleaner and easier to reason about and in come cases means that we can return multiple errors. As a bonus some of the Json reads code becomes a little easier to reason about as well.

This adds an API endpoint to which YAML can be sent for validation and some very rudimentary web pages that let you do the same (your eyes will hurt!)