When parsing the YAML check for extra fields
If a field is provided that we are not using then this is probably a bad thing. In many cases it will be because someone has made a typo (e.g. they may have used `template` instead of `templates`). This is another opportunity to fail fast to make life easier for users.