Further improve parameter validation
I had been overly enthusiastic about parameter validation and it turned out that there are some cases in which no default is specified and that's still OK (i.e. the parameter has no default but is optional). One of these cases is when there are two parameters and you need to set one and only one of these.

This adds an optional field and also enhances the withDefault functionality to provide the target - this means that we can shift more default logic into the parameters and make fewer params optional when really they default from stack or some such deploy time data.