Extends deployment type documentation to cover actions
This extends the deployment type documentation that currently covers parameters to also cover actions. This has been implemented in exactly the same way as parameters were. Sadly this has not yet stretched to automatic documentation of tasks and I suspect this would be difficult, especially for actions that have parameters without any default.

The overall documentation has also been improved as a result.