Autoscaling: wait for stabilization rather than initial single check
Riff-raff currently executes a single check to verify the ASG and ELB
have the same number of instances.
In case the ASG is currently scaling up/down when ASG deploy is being started,
the check fails and the deploy is terminated.
From now on Riff-raff will wait for the ASG to stabilize (for a short
period of time) before to fails the deploy rather than making a single
check