decorate log line with extra information
Adds a deploy record's metadata as markers to the log line. This is primarily to identify which branch is being deployed to help understand our progress of moving from `master` to `main`.

Logging this information as markers would mean we can create graphs within the Central ELK stack: a graph of branch name used over time.