Update the docs for aws-s3 param publicReadAcl
We ran into an issue using the aws-s3 deployment type where the default
value for publicReadAcl (true) doesn't play nicely with the default
policy for new buckets, which is to restrict all public access. Update
the docs to specify that if your bucket blocks public access then you
should set this param to false.