Stream the S3 copy rather than buffer entirely
When migrating from AWS SDK v1 to v2 the S3 Upload was changed to buffer the entire file being uploaded from bucket A to B in memory. This created extra memory pressure so this change switches back to streaming with a 1Mb buffer.

Due to the way that this is now implemented, the get and put requests must be on different threads. In order to facilitate this we add an isolated execution context for IO which is now passed around so tasks can use that when appropriate.