chore: Add `allowlist` field to DynamoDb
We want to rename "whitelist" to "allowlist" because language matters.

In order to do this safely, we need to perform a three step dance:
  1. Add the new field
  2. Read the new field
  3. Remove the old field

This ensures we can continue to read the data at all times,
at the cost of three separate changes.

This is the first step.

It is done in a hacky way by mutating the DynamoDb table in the GET /deployment/restrictions request.
It's hacky as it makes this route, temporarily, non-idempotent.