Remove Guardian Management deps
These dependencies are very old, very little used outsideo of RR.
They're also difficult to update to Scala 2.13+. As such, we are
removing them for security and maintenance reasons.

Switch-related code has been partially inlined from the library
as part of this to ease the transition.

Also, it should be noted that a fuller switch solution is useful
here, but beyond the scope of this immediate change.