fix!: Set `publicReadAcl` to `false` by default
Following the principle of least privilege, set the `publicAcl` to false by default.

This also fits with AWS's own docs (https://docs.aws.amazon.com/AmazonS3/latest/userguide/acl-overview.html):

> A majority of modern use cases in Amazon S3 no longer require the use of ACLs, and we recommend that you disable ACLs except in unusual circumstances where you need to control access for each object individually.

See https://github.com/guardian/riff-raff/pull/141 for a discussion about this - the overall conclusion was to have a default of `false` too.

BREAKING CHANGE: You'll now have to explicitly opt-in to this setting.

This change would break projects that rely on the default being `true` and have not explicitly set the value in their `riff-raff.yaml`.
It's pretty difficult to identify these projects - there are a lot or projects that explicitly set it to false though!
I'd suggest we address any issues when they come up.