Upgrade Scala, Play, SBT
Scala -> 2.13 (previously 2.12)
Play  -> 2.8 (2.6)
SBT   -> 1.5.6 (0.13)

This was a bit of a gigantic upgrade due to dependency binary
incompatibilities (dependency coupling) and general breaking
changes.

Hopefully in future we can do smaller and more frequent updates.

To break down the changes:

1. Removal of some niche dependencies

The list is: diffson-play-json, scala-ssh, henkan-convert.

These were either unused or used only a little.

2. .mapValues -> .view.mapValues(...).toMap

https://docs.scala-lang.org/overviews/core/collections-migration-213.html#what-are-the-breaking-changes

3. other Scala 2.12 and 2.13 changes

* .par now requires a special import
* Seq is now core.immutable.Seq
* import precendence and shadowing seems to have changed (see:
  https://github.com/scala/scala/pull/6589), which caused issues
  as task and deployment_types packages share many names
* requirement to drop parens when a method (typically controller
  method) was defined without them
* etc.

4. library breaking changes

For example, in the Guardian's own Play Secret and GoogleAuth
libraries, or in the Boostrap ('b3') library, especially in
how forms are defined.

5. SBT changes

The deprecation of 'in' for '/' instead.