fix: Use raw YAML to determine Anghammarad targets
Currently, we validate the riff-raff.yaml before identifying targets for an Anghammarad notification.

This is fine until we change the rules for what makes a valid riff-raff.yaml file!

In this change we remove the validation step and instead use the raw YAML file.

See https://github.com/guardian/riff-raff/pull/665.