Attempt fix of scheduled deploys issue
The error seen is:

software.amazon.awssdk.core.exception.SdkClientException Unable to marshall request to JSON: Can not write a field name, expecting a value

And it is kicked off here:

magenta.tasks.UpdateS3Lambda.$anonfun$execute$7(tasks.scala:263)

though ultimately is thrown from the AWS SDK.

It's a bit unclear why this has broken. There are not many similar
reports, but an example is:

https://github.com/aws/aws-sdk-java-v2/issues/2875

As the inputs to the call look okay (we log them in the build) for
now bumping the library makes sense as a first effort at a fix.