fix: Poll CloudFormation events only when necessary
The `cloud-formation` deployment type in Riff-Raff has a number of steps:
  1. Set the update policy on the CloudFormation stack to guard against deleting stateful resources
  2. Create a ChangeSet
  3. Execute the ChangeSet
  4. Watch the CloudFormation event log to determine when 3 has finished
  5. Revert 1

Currently in step 4, we're only looking at the _first_ page of CloudFormation events and we expect to find an `UPDATE_IN_PROGRESS` or `CREATE_IN_PROGRESS` event in the list.
When we don't, the deploy currently fails with `No events found at all for stack...`.

This happens if a CloudFormation update generates _lots_ of items in the event log, causing the `UPDATE_IN_PROGRESS` or `CREATE_IN_PROGRESS` events off the first page.

In this change we update step 4, exiting early if the change set created in step 2 will not yield changes to the stack.
Meaning we only look at the event log when necessary. This logic already happens for step 3.

We subclass `CheckUpdateEventsTask` into `CheckUpdateEventsForChangeSetTask` for use in the CloudFormation deployment type for ease (this is a pragmatic change).
`CheckUpdateEventsForChangeSetTask` will exit early if there the ChangeSet created at the start of the deployment did not yield any changes.