Fix Google Cloud Upload Task
Was failing with:

java.lang.NoClassDefFoundError com/google/api/client/json/jackson2/JacksonFactory

This looks like it was caused by multiple versions of some of the
Google client libraries. The evicted version contained this class
but the preferred version did not.

play-googleauth pulls in a newer one here than our storage library
did so I've switched the storage dependency for the newer version.