fix: Poll CloudFormation events within `ExecuteChangeSetTask`
In #671 we made some changes exiting early if a ChangeSet does not yield any changes.
We accomplished this by calling `DescribeChangeSet` within `CheckUpdateEventsForChangeSetTask`.
This happens after `ExecuteChangeSetTask` but it is too late!

It appears a ChangeSet cannot be described by name once it has been executed; it can only be described using the ARN.

In this change, we perform the CloudFormation event log polling within `ExecuteChangeSetTask` as opposed to after it.
This is necessary because we don't (yet) have a mechanism to conditionally perform a `Task`.
That is, it is not possible to call `CheckUpdateEventsForChangeSetTask` if and only if `ExecuteChangeSetTask` generated changes to the CloudFormation stack.

It is also not possible to pass data between tasks. We only know the ChangeSet name within the task orchestrator,
therefore `DescribeChangeSet` cannot be performed within `CheckUpdateEventsForChangeSetTask` as it is after `ExecuteChangeSetTask`.
The only option is to combine the two (and consequently remove `CheckUpdateEventsForChangeSetTask` as it's now unused).

By performing the event log polling within `ExecuteChangeSetTask`, we're also able to observe the execution status of the ChangeSet.
Now, in addition to waiting for the event log to contain a "complete" message, we also wait until the ChangeSet has successfully completed.

Co-authored-by: Michael McNamara <michaelwmcnamara@users.noreply.github.com>