refactor: Remove `docker-testkit-*` dependencies in favour of `docker-compose`
Currently, `docker-testkit-scalatest` and `docker-testkit-impl-spotify` is used to orchestrate a Docker container running Postgres in test.

This is undesired for a number of reasons:
  - We have to manually keep the content of `docker-compose.yml` and `DockerPostgresService.scala` in sync
  - `docker-testkit-*` does not work on Apple M1 machines (`java.lang.UnsatisfiedLinkError` is seen)
  - It's an extra set of dependencies
  - It creates a coupling between tests and Docker

In this change, we remove these dependencies in favour of running `docker-compose` commands:
  - `docker-compose up -d` during test setup to start containers as described in `docker-compose.yml` (Postgres)
  - `docker-compose rm --stop --force` during test cleanup to stop and remove the aforementioned containers

The latter happens when running tests locally too; all content created in the database during local development will be removed after running the test suite locally.