refactor: Terminate instance directly rather than via ASG
By directly terminating itself, we can issue a tighter IAM policy.

From allowing termination of any instance in the ASG:

```yaml
- Effect: Allow
  Action: autoscaling:TerminateInstanceInAutoScalingGroup
  Resource: !Sub arn:aws:autoscaling:${AWS::Region}:${AWS::AccountId}:autoScalingGroup:*:autoScalingGroupName/${AWS::StackName}-AutoscalingGroup-*
```

To allowing termination of only itself:

```yaml
- Effect: Allow
  Action: ec2:TerminateInstances
  Resource: '*'
  Condition:
    StringEquals:
      aws:ARN: ${ec2:SourceInstanceARN}
```

In reality, this makes little difference as Riff-Raff operates as a singleton.
However is best to follow principles of least privilege.

The changes to the length of downtime is minimal, 5 seconds more, based on one test.