feat(autoscaling): Enable deployment to a new ASG
As part of the pre-flight checks for an `autoscaling` deployment, Riff-Raff will attempt to lookup the ASG.

This is fine if the ASG already exists. However, if we want to create a new stack (from nothing) we cannot
use Riff-Raff for the initial creation and have to resort to going directly to AWS.

In this change, we change the rules.

If an ASG cannot be located and the update strategy is 'Dangerous', then we allow the deployment to continue.
This allows us to use Riff-Raff to deploy a new stack from scratch.

If an ASG cannot be located and the update strategy is 'MostlyHarmless', we continue to fail, but provide a hint to the 'Dangerous' option.