Delete obsolate files in Google Cloud Platform deploys (#740)
Alters the GCP deployment type to delete files which are not included in the current deploy. This is an optional configuration. The new configuration will take a list of folders ( relative to the GCS deploy root ) and a list of filetypes to purge. If these are absent from the configuration the behaviour is as previously