refactor: Simplify JSON for stack policy
JSON formatting is hard...

We currently render a JSON array as:

```json
{
    "ResourceTypes": [
        "One",
"Two",
"Three"
    ]
}
```

That is, without indentation.

It should be:

```json
{
    "ResourceTypes": [
        "One",
        "Two",
        "Three"
    ]
}
```

However, there isn't much need to pretty-print this JSON blob as machines are the only consumers.
In this change, we render a JSON array as a simple comma separated value:

```json
{
    "ResourceTypes": [
        "One","Two","Three"
    ]
}
```

This is functionally equivalent, and makes the tests simpler to read.