Record build tool
Adds the buildTool as a tag on Cloudformation deploys. This will
help us to track usage of build tools for services. The particular
focus is @guardian/actions-riff-raff but the solution is general.

Note, as part of this, the reporter error for missing vcsUrl has
been removed - looking more closely at the code, I believe deploys
should already fail if this is not present so the warning here is
redundant.