Add SSM bucket parameters to S3 deployment type
Best practice is now to avoid explicit bucket names, which should
be considered private, in riff-raff.yaml files. SSM parameters
are our mechanism to achieve this and S3 was a notable exception
to this until now.

Note, as part of this change the 'pathPrefixResource' parameter
has been removed as part of a general deprecation and removal of
custom resources. There is a single user of this parameter and
they have been informed.

We will remove the related 'bucketResource' in a follow-up PR once
the (again single) user has migrated successfully to the
'ssmKeyStageParam' param for their affected S3 deployment.