fix: Fallback to `amiEncrypted` when `amiParametersToTags/Encrypted` is absent
The changes in #991 has caused https://riffraff.gutools.co.uk/deployment/view/58385bf3-b10f-47b3-afac-d9e53e1a7278 to fail.

This is because the `riff-raff.yaml` is:

```yaml
regions:
- eu-west-1
stacks:
- ophan
deployments:
  big-loader-es7:
    type: autoscaling
    dependencies:
    - update-ami
  update-ami:
      type: ami-cloudformation-parameter
      app: big-loader-es7
      parameters:
        amiEncrypted: true
        amiParametersToTags:
          ImageId:
            Recipe: ophan-ubuntu-bionic-ARM
            AmigoStage: PROD
            BuiltBy: amigo
```

That is, `amiParametersToTags/Encrypted` is not set, so `Encrypted=false` is used.
However `Encrypted=true` should actually be used.

In this change we fallback to `amiEncrypted` if `amiParametersToTags/Encrypted` is absent.

Ideally, support for `amiEncrypted` would be dropped, however multiple projects are using this parameter already.