Don't expose branch regexs for CD, just names
Looking at projects with CD configured, it seems that the only type of
configuration is a branch name, which gets convered to a regex (by
adding start and end anchors). Lets simplify the UI and make it clear
you can enter a branch name, and expose that name back to users, rather
than the constructed regex. I've gotten confused in the past and added
the ^ and $ manually, which isn't necessary (I see a few instances of
this in the CD configurations, so I don't think I'm the only one!)