Fix broken tests
I found it slightly hard to see at a glance which of the shared configs
and builds were needed in which tests, so I've refactored to move the
test setup into the specific test which requires it. It's a bit more
verbose but I think it's easier to follow. I've used the Arrange Act
Assert pattern to hopefully make the individual tests clearer.