Order builds by start time
This was prompted by a recent bug where a newer build had a lower
ID (which can happen with github actions): continuous deployment
deployed the older build twice, and didn’t deploy the newer build at
all.

The implicit ordering on `CIBuild` is the cause: it’s used in
`ContinuousDeployment.buildCandidates` by `GreatestSoFar` to order the
builds produced by the poller.

I’ve also changed the ordering used in `Builds.successfulBuilds` to
match: I don’t think this is currently used anywhere, but it’s good to
be consistent in case it’s used in the future.

Writing a meaningful test of `ContinuousDeployment.buildCandidates` to
test the ordering seems like a lot of effort, so I haven’t done it.
Hopefully the more basic test added here is useful.