Don't grant AUTHORIZATION when creating service_catalogue schema.
There seems to be an issue in RDS where when creating a schema with AUTHORIZATION the user creating the schema must have the role that the schema authorization is being applied to (riffraff needs service_catalogue in this case).

Workaround this by removing AUTHORIZATION and granting USAGE to service_catalogue after the fact.

https://stackoverflow.com/questions/26684643/error-must-be-member-of-role-when-creating-schema-in-postgresql