fix: Report build number correctly
We're currently using the environment variable `GITHUB_RUN_NUMBER` to determine the build number for build info.

`GITHUB_RUN_NUMBER` is not offset by the number of builds of this project previously performed by TeamCity.

This change creates a new environment variable `BUILD_NUMBER` which is `GITHUB_RUN_NUMBER` + the TeamCity offset.