Give WaitForCullToComplete up-to-date ASG info!
Jacob Winch pointed out that `WaitForCullToComplete` is a repeating check, and
so needs to get up-to-date `AutoScalingGroupInfo` in order for it to know what
instances currently exist and what their state is! I was missing a call to
`ASG.refresh(asg)`.

This is easy to miss in tasks like `WaitForCullToComplete` that extend
`magenta.tasks.PollingCheck`, the
`magenta.tasks.ASGTask.execute(asg: AutoScalingGroup, ...)` method
puts a `AutoScalingGroup` into scope by the method parameter, and it's inevitably
out-out-date... could be something to refactor in a later PR!

We also decided that as polling checks inevitably involve network calls, it makes
sense to put an exception-catching guard in the `magenta.tasks.PollingCheck.check()`
method.