fix(ami-cloudformation-parameter): Perform update within a CFN change set
Perform an AMI update via CFN change set for increased reliability.

When creating a change set, AWS will calculate what, if any, diff there is.
When there is no diff, we will now succeed fast.

This differs from current behaviour where we poll the CFN event log,
even if the CFN update was a no-op.

More specifically, we poll the first page of the CFN event log,
and use a particular status message as an anchor point.
If this anchor point is on the 2+ page, the deployment fails.
This can happen if a CFN change affected lots of resources,
for example a new tag was added to the stack, or an ASG rolling update was performed.