feat(asg-rolling-update): Fail deployments when scaled to >=75% capacity
A request from the MAPI team as part of exploring rolling update deployments,
this change will fail to deploy if a service has scaled to >=75% of capacity.

Previously, we would effectively scale-in the service in this scenario
for the sake of having enough capacity for new instances.
In reality, this is undesired.

See https://github.com/guardian/mobile-apps-api/discussions/3747#discussioncomment-13446295.