Use manual query parameters for correct route
Since the route doesn’t expore the query parameters, I can’t do this in
a nice type-safe Scala way. So this commit writes those query parameters
into the url string manually. This is less durable (i.e., we won’t get a
compile error if the route’s query parameters change in a way that
breaks this), but it doesn’t take as much work as refactoring the route
to expose the query parameters (which I don’t want to do).

(Hopefully it’s relatively unlikely we’ll change these parameters? At
least the breakage will be minor.)