Add an expand-all button to the deploy log
Riff Raff often emits warnings that are very challenging to find in
the deploy log. We add an "expand all" button that ensures the UI is
showing all log entries, which makes it much easier for a human to
quickly scan the output to find warnings (which are highlighted in the
UI).

This isn't a perfect solution to the problem, because it depends on
someone being able to visually parse the error warning. It's an easy
fix though, and very significantly improves these situations.

Addresses https://github.com/guardian/riff-raff/issues/1459, which
provides much more context and detail.