feat: Set `RiffRaffDeploymentId` CFN template parameter (when available)
If a CloudFormation template has a `RiffRaffDeploymentId` parameter,
Riff-Raff will set it to the deployment's GUID.

Paired with https://github.com/guardian/cdk/pull/2723, this improves the DX when using the new CFN-based ASG deployment mechanism.
Previously a redeployment of the same build was a no-op - no instances were cycled.
To force instances to be cycled, a new build was required, which can be slow.
With this change (if `RiffRaffDeploymentId` is also echoed in the user data), a redeployment of the same build cycles instances.

Co-authored-by: Jacob Winch <jacobwinch@users.noreply.github.com>