WIP: reload deploy progress without overwriting
The idea here is to only replace bits that are in progress with new
bits.

I haven’t run this yet and it definitely doesn’t work. There’s no
framing yet like there is in auto-refresh.coffee to pull out the URL
from the DOM. Plus half of it is javascript: will coffeescript tolerate
that? 🤷